@@ -1,5 +1,5 @@
 module OpenProject
   module GwdgProjectGroups
-    VERSION = "6.0.5"
+    VERSION = "6.1.0"
   end
 end
